V4.0

+++ view/js/js.js
@@ -28,7 +28,7 @@ function addSalle() {
 function calculePrixTT() {
   var quantite = document.getElementById("quantite");
   var prix = document.getElementById("prix");
-  var res = parseInt(quantite.value) * parseFloat(prix.value);
+  var res = parseFloat(quantite.value) * parseFloat(prix.value);
   document.getElementById("prixTT").value = res;
 }
